requirements.txt
pika
python-dotenv
Faker
apache-beam[gcp]
pyarrow
pandas
duckdb
streamlit